refactor(node-utils): move createGetSecretsFromEnv to node-utils
