+++ packages/nodeUtils/src/index.js
@@ -15,8 +15,16 @@
 */
 
 import cleanDirectory from './cleanDirectory';
+import createGetSecretsFromEnv from './createGetSecretsFromEnv';
 import getFileExtension, { getFileSubExtension } from './getFileExtension';
 import readFile from './readFile';
 import writeFile from './writeFile';
 
-export { cleanDirectory, getFileExtension, getFileSubExtension, readFile, writeFile };
+export {
+  cleanDirectory,
+  createGetSecretsFromEnv,
+  getFileExtension,
+  getFileSubExtension,
+  readFile,
+  writeFile,
+};
